Except for the screen drivers, currently the drivers work as a substitute to "standard library" functions; The drivers call the library functions or don't function at all. The side effect of this is that drivers are not used as they should be. This should be fixed.
An example of this is the keyboard function. "GetChar" does not interface with the driver to retrieve data (even though it should).

Anywho I implemented the foundation to the addition of such a system. It's pretty much a new driver layer. A barebones implementation is being done for the file system first.